complete wtt display
